Python dependency workarounds
